Clean up init, main menu, and autorun
There was a jumble of jumping around that was confusing in the code.
Organized the tasks into subs to call the one-time init from the top RUN
and auto run (now at 100 - Makefile had it wrong at 1 before, it was
supposed to be 500 before), then have the main menu do the calls to the
top screen draw, UDG load, then bottom screen draw. Now that the UDGs skip
if already loaded, returning to the main menu is not a problem.