Beat on the Makefile some more
The problem is that if you define a target name, you need to give it a
file that it can check to build, and you need to have a target for that
file with dependencies it can check. At least now, the levels tap file
depends on the individual files we cat into it, so it can check if they
are newer than the leles tap file. Also, the dist target should depend on
the levels tap file.